finish annexation of Vietnam, demand Shanghai
Also finished the sphere of South America by joining the Argentine
Confederation's restoration of order to part of Paraguay and calling in
literally the rest of South America to also add Paraguay to my sphere.

Demanded a treaty port from China mostly to fulfill a poorly written
condition later that requires me to have land in the asian continent to
annex the Sulu, but Sighpie reworked the continents so now Indochina and
the East Indies aren't parts of asia, oops.